refactor : 테스트 패키지 수정 및 필요없는 new 연산자 삭제
